eLibrary V2023.11 - Core::Array Constructor<Es...>(Es...) / begin / doAllocate / end / setElement(New) Constructor(intmax_t)(Remove) - Core::ArrayList addElement / doReverse / removeIndex(Arrays::doMove Implementation) Constructor(Array<E>)(New) Constructor<size_t>(std::array) / toSTLArray(Remove) - Core::ArraySet addElement / removeElement(Arrays::doMove Implementation) - Core::Arrays / doCopy / doCopyBackward / doCopyReverse / doMove / doMoveBackward / doMoveReverse(New) - Core::Collection / isEmpty(New) - Core::DoubleLinkedIterator(New) - Core::DoubleLinkedList toArrayList(Remove) - Core::IndexException(New) - Core::List(New) - Core::Map / isContainsKey / isContainsValue(New) - Core::Mathematics doCeil2(getNearestPower2Upper) doFloor2(getNearestPower2Lower) - Core::NtModule(Remove) - Core::NtProcess doCreate(Remove) - Core::Objects doCall / doForward / doMatchExecute / doMatchValue / doMove / doSwap(New) - Core::Queue(New) - Core::Set(New) - Core::SingleLinkedIterator(New) - Core::SingleLinkedList begin / end(New) - Core::SingleLinkedQueue / addBack / addFront / doAssign / doClear / getBack / getFront / isEmpty / removeBack / removeFront / toString(New) - Core::SingleLinkedSet begin / end(New) - Core::Stack(New) - Core::TypeException(New) - Core::WindowsException(New) - IO::Buffer doDiscardMark / doMark / doReset / getBufferMark(Remove) - IO::ByteBuffer doCompact / doCompare(New) - IO::FileDescriptor doOpen(Remove) - IO::FileInputStream Constructor(const String&)(Remove) doOpen<Os...>(const String&, Os...)(doOpen(const String&)) doSeek(long, int)(doSeek(off64_t, int)) - IO::FileOption(New) - IO::FileOutputStream Constructor(const String&, bool=false)(Remove) doFlush(...) doOpen<Os...>(const String&, Os...)(doOpen(const String&, bool)) doSeek(long, int)(doSeek(off64_t, int)) doTruncate(Remove) - IO::Library / doOpen / getSymbol(New) - IO::OutputStream doWrite(uint8_t)(doWrite(int)) - Network::NetworkAddress Constructor(Array<uint8_t>)(New) Constructor(std::array<uint8_t, 4>) / Constructor(std::array<uint8_t, 16>) / Constructor(std::initializer_list<uint8_t>)(Remove) isLoopbackAddress(...) - Network::NetworkSSLDescriptor doRead(void*, int)(doRead(void*, uint8_t)) doWrite(void*, int)(doWrite(void*, uint8_t))

+++ eLibrary/IO/Stream.hpp
@@ -35,7 +35,7 @@ namespace eLibrary::IO {
 
         virtual void doFlush() {}
 
-        virtual void doWrite(int) = 0;
+        virtual void doWrite(uint8_t) = 0;
 
         virtual void doWrite(uint8_t *StreamBuffer, uint32_t StreamOffset, uint32_t StreamSize) {
             if (!isAvailable()) throw IOException(String(u"OutputStream::doWrite(uint8_t*, uint32_t, uint32_t) isAvailable"));
